requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.2
python-dotenv==1.0.0
Werkzeug==2.3.7
SQLAlchemy==2.0.21
Flask-Mail==0.9.1
email-validator==2.1.0.post1 